+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `Text input` component (Orange-OpenSource/ouds-ios#406)
+- `text input` component (Orange-OpenSource/ouds-ios#406)
 - `body moderate` and `label moderate` typography and tokens (Orange-OpenSource/ouds-ios#1003)
 - Management of `Helvetica Neue Arabic` font family (Orange-OpenSource/ouds-ios#1006)
 - UIKit experimental backports for `button`, `tag`, `badge`, `horizontal divider`, `vertical divider`, `link`, `suggestion chip`, `filter chip`, `checkbox`, `checkbox indeterminate`, `checkbox item`, `radio`, `radio item`, `switch` and `switch item` components